+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Craft CMS 3 - Mailjet mailer adapter
 
+## v1.1.0 - 2019-02-27
+
+### Added
+
+- SMS features by Mailjet API
+- System utility to view sent messages
+- Caching for better performance in utility
+
+### Changed
+
+- Fixed checking if Mailjet transport is used in `MailjetAdapter::isUsed()`
+- Separated docs into multiple markdown files
+
 ## v1.0.0 - 2019-02-23
+
+### Added
 
 - Initial release of the Craft CMS 3 plugin for Mailjet